Allow kapp binary path to be configured in e2e tests
This commit introduces the `KAPP_BINARY_PATH` environment varialbe so
that the path to the binary can be configured when running the end to
end suite. If not specified, `KAPP_BINARY_PATH` will default to the
`kapp` that is on the `PATH`. When running `./hack/test-all.sh`, the
`KAPP_BINARY_PATH` will be configured to the binary that is built.

Fixes #55