Make IdentifierResources.Patch method to return Resource with stripped identity annotation
Other methods, like Create and Update, are stripping annotations,
this change makes Patch consistent with the rest.

Other methods force identity annotation to be set on all resources sent to K8S API.
Technically Patch should do the same with patch data, but it is tricky as there are multiple
patch formats supported by K8S and analysing and amending them to force identity annotation
to be present doesn't worth the effort.