Preserve status for all resources on update (#411)
* Preserve status for all resources on update

Update the default config to ignore 'new' status, and instead preserve
the a resource's 'existing' status.

Resources that define a status will no longer try to update the server
to make that status empty. Status may not be defined when creating a
resource (this is enforced by k8s) and should not be defined on update,
as this state should be owned by the controller for that resource.

Many tools will include "empty" scaffolding for a resource when
unmarshaling a struct. These values will now be ignored for status. For
example, controllergen includes a complex empty object for the status
of CustomResourceDefinitions that kapp will now ignore.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>

* prune non-required fields

Signed-off-by: Scott Andrews <andrewssc@vmware.com>