Ignore rules for aggregate ClusterRoles
Aggregate ClusterRole have their rules filled in by the control plane at
runtime. There is no need for kapp to manage this field as it will only
create churn that the control plane has to then undo.

By adding a default rebase rule for ClusterRoles that have an
aggregationRule field set, we can ignore the rules on update.

Refs https://kubernetes.io/docs/reference/access-authn-authz/rbac/#aggregated-clusterroles

Signed-off-by: Scott Andrews <andrewssc@vmware.com>