Add fs.FS support to deploy, inspect, diff (#799)
* Flags: use struct fields as flag defaults

Use struct fields as flag defaults. If the flag structs are being filled
in programmatically, the values are cleared during flag parsing if the
flag defaults differ from the current struct values, which is
undesirable.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>

* LocalFileSource: support a custom fs.FS

Add fs.FS support to LocalFileSource, FileResource, and the commands
that use them (deploy, diff, inspect). This is useful so these commands
can be invoked programmatically.

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>

* code review

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>

* code review

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>

---------

Signed-off-by: Andy Goldstein <andy.goldstein@redhat.com>